requirements.txt
beautifulsoup4==4.9.1
python-dotenv==0.13.0
selenium>=4.18.1
webdriver-manager>=4.0.2